refactor: 요약 레포트를 받아오는 훅과 받아온 데이터를 가공하는 훅을 분리 적용
